move to app guard from main app
