+++ src/app/MainApp.tsx
@@ -16,7 +16,6 @@ import { setupGlobalKeyboardListeners } from '@/shared/utils/keyboard';
 import type { FileAttachment } from '../../worker/types';
 import { getConversationEndpoint, getConversationsEndpoint, getCurrentConversationEndpoint } from '@/config/api';
 import { linkConversationToUser } from '@/shared/lib/apiClient';
-import { getClient } from '@/shared/lib/authClient';
 import { useNavigation } from '@/shared/utils/navigation';
 import {
   BanknotesIcon,
@@ -344,7 +343,6 @@ export function MainApp({
   // Removed unused submitUpgrade
   const { showError } = useToastContext();
   const showErrorRef = useRef(showError);
-  const onboardingCheckRef = useRef(false);
   const practiceWelcomeCheckRef = useRef(false);
   const isSelectingRef = useRef(false);
   useEffect(() => {
@@ -354,8 +352,7 @@ export function MainApp({
     currentPractice,
     acceptMatter,
     rejectMatter,
-    updateMatterStatus,
-    refetch: refetchPractices
+    updateMatterStatus
   } = usePracticeManagement({
     autoFetchPractices: workspace !== 'public',
     fetchInvitations: workspace !== 'public'
@@ -374,31 +371,7 @@ export function MainApp({
     void fetchPracticeDetails();
   }, [fetchPracticeDetails, hasPracticeDetails, practiceDetailsId]);
 
-  // Handle subscription success - refresh session and practice list after Stripe checkout
-  const subscriptionSuccessHandledRef = useRef(false);
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    if (subscriptionSuccessHandledRef.current) return;
-    
-    const urlParams = new URLSearchParams(window.location.search);
-    if (urlParams.get('subscription') !== 'success') return;
-    
-    subscriptionSuccessHandledRef.current = true;
-    
-    if (import.meta.env.DEV) {
-      console.debug('[SUBSCRIPTION][SUCCESS] Refreshing session and practice list after checkout');
-    }
-    
-    // Refetch practices to get the newly created practice
-    // The session should already be refreshed by Better Auth after the Stripe redirect
-    void refetchPractices().then(() => {
-      if (import.meta.env.DEV) {
-        console.debug('[SUBSCRIPTION][SUCCESS] Practices refetched successfully');
-      }
-    }).catch((error) => {
-      console.error('[SUBSCRIPTION][SUCCESS] Failed to refetch practices:', error);
-    });
-  }, [refetchPractices]);
+
 
   const handleMessageError = useCallback((error: string | Error) => {
     console.error('Message handling error:', error);
@@ -677,69 +650,7 @@ export function MainApp({
   // Clients chat with practices via widget (practiceId from URL), not their own practice.
 
 
-  useEffect(() => {
-    if (sessionIsPending || isAnonymous || !session?.user?.id) {
-      if (import.meta.env.DEV) {
-        console.debug('[ONBOARDING][SKIP] session pending or anonymous', {
-          sessionIsPending,
-          isAnonymous,
-          hasUser: Boolean(session?.user?.id)
-        });
-      }
-      return;
-    }
-    
-    // Skip onboarding check if user is returning from successful subscription
-    // Check this BEFORE the onboardingCheckRef guard so it's evaluated on every URL change
-    if (typeof window !== 'undefined') {
-      const urlParams = new URLSearchParams(window.location.search);
-      if (urlParams.get('subscription') === 'success') {
-        if (import.meta.env.DEV) {
-          console.debug('[ONBOARDING][SKIP] returning from successful subscription', {
-            url: window.location.href
-          });
-        }
-        return;
-      }
-    }
-    
-    if (onboardingCheckRef.current) return;
-    onboardingCheckRef.current = true;
 
-    const checkOnboarding = async () => {
-      try {
-        if (import.meta.env.DEV) {
-          console.debug('[ONBOARDING][CHECK] fetching preferences');
-        }
-        const prefs = await getPreferencesCategory<OnboardingPreferences>('onboarding');
-        const needsOnboarding = prefs?.completed !== true;
-
-        if (import.meta.env.DEV) {
-          console.debug('[ONBOARDING][CHECK] preferences', {
-            completed: prefs?.completed ?? null,
-            birthday: prefs?.birthday ? '[REDACTED]' : null,
-            primary_use_case: prefs?.primary_use_case ?? null,
-            use_case_additional_info: prefs?.use_case_additional_info ? '[REDACTED]' : null
-          });
-        }
-
-        if (needsOnboarding) {
-          if (import.meta.env.DEV) {
-            console.debug('[ONBOARDING][REDIRECT] redirecting to /auth?mode=signin&onboarding=true', {
-              reason: 'completed flag not true',
-              userId: session?.user?.id ?? null
-            });
-          }
-          window.location.href = '/auth?mode=signin&onboarding=true';
-        }
-      } catch (error) {
-        console.warn('[ONBOARDING][CHECK] preferences fetch failed:', error);
-        onboardingCheckRef.current = false;
-      }
-    };
-
-    void checkOnboarding();
-  }, [isAnonymous, session?.user?.id, sessionIsPending]);
 
   // Check if we should show practice welcome modal
   useEffect(() => {
